# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ReMynd Assessment Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: phase
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get case details
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseRequest"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}/advance:
    post:
      operationId: advanceCasePhase
      tags: [cases]
      summary: Advance case to next phase
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}/intake-analysis:
    post:
      operationId: analyzeIntake
      tags: [cases]
      summary: Run AI intake analysis
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Intake analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeAnalysis"

  # ─── Assessment Tools ─────────────────────────────────────────────────────
...
  /cases/{caseId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment(s) for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /cases/{caseId}/assignments/{assignmentId}:
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment status
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentRequest"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  # ─── External Respondent ─────────────────────────────────────────────────
  /external/form/{token}:
    get:
      operationId: getExternalForm
      tags: [external]
      summary: Get form for external respondent (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Form data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalForm"
        "404":
          description: Token not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /external/form/{token}/submit:
    post:
      operationId: submitExternalForm
      tags: [external]
      summary: Submit form response (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFormRequest"
      responses:
        "200":
          description: Submission accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitFormResponse"

  # ─── Self-Report Admin ────────────────────────────────────────────────────
...
  /cases/{caseId}/report:
    get:
      operationId: getCaseReport
      tags: [reports]
      summary: Get the report for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: No report yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cases/{caseId}/report/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: AI-generate report
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /cases/{caseId}/report/update:
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Admin edits to the report
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportRequest"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /cases/{caseId}/report/approve:
    post:
      operationId: approveReport
      tags: [reports]
      summary: Admin approves report
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ─── Users ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Case:
      type: object
      properties:
        id:
          type: string
        studentName:
          type: string
        dob:
          type: string
        school:
          type: string
        grade:
          type: string
        languagePreference:
          type: string
          enum: [english, mandarin, cantonese]
        referralReason:
          type: string
        caseStatus:
          type: string
          enum: [active, on_hold, completed, archived]
        currentPhase:
          type: string
          enum: [pre_commitment, intake, setup, forms, assessment, scoring, report, debrief, complete]
        progressPercentage:
          type: integer
          minimum: 0
          maximum: 100
        riskLevel:
          type: string
          enum: [low, moderate, high]
          nullable: true
        assignedLeadId:
          type: string
          nullable: true
        assignedPsychId:
          type: string
          nullable: true
        parentName:
          type: string
          nullable: true
        parentEmail:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        consentObtained:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - studentName
        - dob
        - school
        - languagePreference
        - caseStatus
        - currentPhase
        - progressPercentage
        - consentObtained
        - createdAt
        - updatedAt
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/Case"
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/Assignment"
            scores:
              type: array
              items:
                $ref: "#/components/schemas/Score"
            intakeData:
              type: object
              nullable: true
            intakeAnalysis:
              $ref: "#/components/schemas/IntakeAnalysis"
              nullable: true
    CreateCaseRequest:
      type: object
      properties:
        studentName:
          type: string
        dob:
          type: string
        school:
          type: string
        grade:
          type: string
        languagePreference:
          type: string
          enum: [english, mandarin, cantonese]
        referralReason:
          type: string
        parentName:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        assignedLeadId:
          type: string
        assignedPsychId:
          type: string
      required: [studentName, dob, school, languagePreference, referralReason]
    UpdateCaseRequest:
      type: object
      properties:
        studentName:
          type: string
        school:
          type: string
        grade:
          type: string
        languagePreference:
          type: string
          enum: [english, mandarin, cantonese]
        caseStatus:
          type: string
          enum: [active, on_hold, completed, archived]
        currentPhase:
          type: string
          enum: [pre_commitment, intake, setup, forms, assessment, scoring, report, debrief, complete]
        riskLevel:
          type: string
          enum: [low, moderate, high]
        parentName:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        assignedLeadId:
          type: string
        assignedPsychId:
          type: string
        consentObtained:
          type: boolean
    IntakeAnalysis:
      type: object
      properties:
        recommendedDomains:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [low, moderate, high]
        summary:
          type: string
        flags:
          type: array
          items:
            type: string
      required: [recommendedDomains, riskLevel, summary]

    # ─── Assessment Tool ──────────────────────────────────────────────────────
...
    Assignment:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        toolId:
          type: string
        toolName:
          type: string
        respondentType:
          type: string
          enum: [parent, teacher1, teacher2, student, self]
        respondentLabel:
          type: string
        assignedToName:
          type: string
          nullable: true
        assignedToEmail:
          type: string
          nullable: true
        uniqueToken:
          type: string
        uniqueLink:
          type: string
        qrCodeData:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, completed, overdue]
        dueDate:
          type: string
          format: date-time
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - toolId
        - toolName
        - respondentType
        - respondentLabel
        - uniqueToken
        - uniqueLink
        - qrCodeData
        - status
        - createdAt
    CreateAssignmentRequest:
      type: object
      properties:
        toolId:
          type: string
        respondentType:
          type: string
          enum: [parent, teacher1, teacher2, student, self]
        respondentLabel:
          type: string
        assignedToName:
          type: string
        assignedToEmail:
          type: string
        dueDate:
          type: string
          format: date-time
      required: [toolId, respondentType, respondentLabel]
    UpdateAssignmentRequest:
      type: object
      properties:
        status:
          type: string
          enum: [not_started, in_progress, completed, overdue]
        assignedToName:
          type: string
        assignedToEmail:
          type: string
        dueDate:
          type: string
          format: date-time

    # ─── Response ─────────────────────────────────────────────────────────────
...
    ExternalForm:
      type: object
      properties:
        assignmentId:
          type: string
        toolName:
          type: string
        respondentLabel:
          type: string
        studentName:
          type: string
        language:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/FormQuestion"
        alreadySubmitted:
          type: boolean
      required: [assignmentId, toolName, respondentLabel, studentName, language, questions, alreadySubmitted]
    FormQuestion:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        textMandarin:
          type: string
          nullable: true
        textCantonese:
          type: string
          nullable: true
        type:
          type: string
          enum: [likert, yes_no, text, scale]
        options:
          type: array
          items:
            type: string
          nullable: true
        domain:
          type: string
      required: [id, text, type, domain]
    SubmitFormRequest:
      type: object
      properties:
        answers:
          type: object
          additionalProperties: true
        language:
          type: string
      required: [answers, language]
    SubmitFormResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Score:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        toolId:
          type: string
        toolName:
          type: string
        respondentType:
          type: string
        rawScore:
          type: number
          nullable: true
        domainScores:
          type: object
          additionalProperties:
            type: number
        normalizedScores:
          type: object
          additionalProperties:
            type: number
        agreementIndex:
          type: number
          nullable: true
        hasHighDiscrepancy:
          type: boolean
        isManual:
          type: boolean
        notes:
          type: string
          nullable: true
        generatedAt:
          type: string
          format: date-time
      required:
        - id
        - caseId
        - toolId
        - toolName
        - respondentType
        - domainScores
        - normalizedScores
        - hasHighDiscrepancy
        - isManual
        - generatedAt
...
    Report:
      type: object
      properties:
        id:
          type: string
        caseId:
          type: string
        status:
          type: string
          enum: [draft, approved, exported]
        backgroundSummary:
          type: string
        domainAnalysis:
          type: string
        strengths:
          type: string
        areasOfConcern:
          type: string
        crossSettingComparison:
          type: string
        recommendations:
          type: string
        adminNotes:
          type: string
          nullable: true
        generatedAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - caseId
        - status
        - backgroundSummary
        - domainAnalysis
        - strengths
        - areasOfConcern
        - crossSettingComparison
        - recommendations
        - generatedAt
    UpdateReportRequest:
      type: object
      properties:
        backgroundSummary:
          type: string
        domainAnalysis:
          type: string
        strengths:
          type: string
        areasOfConcern:
          type: string
        crossSettingComparison:
          type: string
        recommendations:
          type: string
        adminNotes:
          type: string

    # ─── Dashboard ────────────────────────────────────────────────────────────
...