# lib/api-zod/src/generated/types/submitFormRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ReMynd Assessment Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitFormRequestAnswers } from "./submitFormRequestAnswers";

export interface SubmitFormRequest {
  answers: SubmitFormRequestAnswers;
  language: string;
}